Add ability to set modes for all apps in a single POST
Currently, the UI form to configure modes per app fires one POST per application
in parallel, (which can be up to ~15), which is leading to MySQL deadlocks.

Avoid this by adding the functionality to set the modes for all of the apps in
just one POST request, as well as supporting the older behavior.