Make the ldap_count metric from sync-targets not get reset
Currently, the ldap_count metric will be emitted as non-zero once
every hour during the sync() loop, and then it will be reset back
to zero.

Instead of that, initially emit the metric as zero, and after that
continually emit the last value we received. This will make looking
at graphs easier to clean changes in the employee count over time.

This is also the old behavior that disappeared several months back.