Modify how the create incident from email functionality works slightly
Instead of checking the "from" header, check the "to" header.

The way this works, is someone emails a special email alias, and then that
email address reroutes the email to iris, *preserving* the _to_ header, and not
rewriting the from header.

Make this functionality work with that setup.

Also, because we're no longer operating on the "from" header, remove
some of the checks against the from header being an employee email,
because it is going to be