Fix team/list names being accepted for OOB user target
Currently, if you send iris an OOB message with the role
being "user" and the target name being the name of a valid mailing
list or team, it will accept the message, and then cause sender to fail
as it won't be able to assign a contact to the message.

Fix that, so instead the request will fail to begin with instead of
causing sender errors.

Also remove some other KeyError's