pep-3110: Fix deprecated exception syntax
PEP-3110 documents the advised syntax changes to exception handling.
Versions of python from 2.6 onwards have supported this syntax (in
preparation for migrations to Python 3.x).

This sereis of changes ensures that all exception handling in Iris is
compliant with this grammar in an effort to make the codebase python 3.x
compatible.