Add support for multi-recipient email notifications
This way, batch emails can be sent without hitting the API each
time. Instead, we send a single email with a list of recipients.
Note that this is currently supported only for out-of-band notifications
that specifically specify email as a contact mode, and does
not support priorities (where messages may be split across many
different contact modes).