Make application name in hook incident data support overwritten app name
When creating test incidents within the API, the application given to the
incident creation hooks is "iris" rather than the application being used.

Fix this such that the hooks receive the overwritten application name.

Also: move definition of `app` to the parent block in the function so it
is more obvious this variable is used later on.