Don't prepend https:// to proxy hostnames (#668)
Historically, this approach has always worked because http forward proxies
generally only listen on http:// (not https://) and urllib3 has not supported
connecting to a http proxy via https:// so it has always ignored the scheme.

However, as of urllib3 >= 1.26 or so, urllib3 does support and attempt connecting
to proxies via https:// (if this schem is provided) and it raises an exception
if the proxy only listens on http://

Fix this by no longer enforcing a http:// prefix to proxy hostnames. If the user
desires connecting to a https:// proxy, this prefix can be provided within Iris's
configuration.